Remove invariantMonoidal instances for most kernel classes and add only invariant instanes
